requirements.txt
playwright>=1.40.0
pytest>=7.4.0
pytest-playwright>=0.4.3
pytest-html>=4.1.0
python-dotenv>=1.0.0
